requirements.txt
aiogram==2.23.1
psycopg2==2.9.5
psycopg2_binary==2.9.3
requests~=2.28.1
python-dotenv~=0.21.0